Navigationsmenyn fungerar
Fick till sist NavigationActivity att fungera som den ska.
